requirements.txt
Keras==2.4.3
numpy==1.19.2
streamlit==0.79.0
face_recognition==1.3.0
pandas==1.2.0
opencv_python==4.5.1.48
Pillow==8.2.0
